[#936] Extract display_profile package (item 8)
Split the platform-specific display ICC profile get/set functions out of
icc_profile/__init__.py into DisplayCAL/icc_profile/display_profile/:
windows.py (registry/WCS-backed get/set/unset), macos.py (AppleScript-backed
get), linux.py (colord/UCMM/XrandR/xprop-backed get). display_profile/__init__.py
holds the get_display_profile() dispatcher that picks the right backend for
sys.platform, and icc_profile/__init__.py re-exports everything so external
consumers are unaffected.

tests/test_icc_profile_hdr_filtering.py monkeypatches winreg/ICCPROFILES as
module attributes on the icc_profile module to drive _winreg_get_display_profile(s);
since those functions now live in display_profile/windows.py, the test's import
was repointed to that module so the patched attributes land in the same globals
the functions actually read from.
