+++ DisplayCAL/icc_profile/__init__.py
@@ -9,20 +9,15 @@ from __future__ import annotations
 
 import binascii
 import contextlib
-import ctypes
 import datetime
 import functools
-import json
 import math
 import operator
 import os
 import pathlib
-import platform
 import re
 import struct
-import subprocess as sp
 import sys
-import warnings
 from hashlib import md5
 from typing import (
     TYPE_CHECKING,
@@ -34,71 +29,9 @@ from typing import (
 )
 from weakref import WeakValueDictionary
 
-if sys.platform == "win32":
-    import winreg
-
-    try:
-        import win32api  # noqa: F401  # availability checked via sys.modules below
-        import win32gui
-    except ImportError:
-        pass
-
-try:
-    from DisplayCAL import colord
-except ImportError:
-
-    class Colord:
-        """Dummy class for colord support."""
-
-        Colord = None
-
-        def quirk_manufacturer(self, manufacturer: str) -> str:
-            """Quirk the manufacturer name.
-
-            Args:
-                manufacturer (str): The manufacturer name to quirk.
-
-            Returns:
-                str: The quirked manufacturer name.
-            """
-            return manufacturer
-
-        def which(self, executable: str, paths: None | list[str] = None) -> None | str:
-            """Check if an executable is available in the system paths.
-
-            Args:
-                executable (str): The name of the executable to check.
-                paths (None | list[str], optional): List of paths to search for
-                    the executable. If None, uses the system PATH.
-
-            Returns:
-                None | str: The full path to the executable if found, else None.
-            """
-            return
-
-    colord = Colord()
 from DisplayCAL import colormath, edid
 from DisplayCAL.defaultpaths import ICCPROFILES, ICCPROFILES_HOME
 from DisplayCAL.options import TEST_INPUT_CURVE_CLIPPING
-from DisplayCAL.util_list import intlist
-
-if sys.platform not in ("darwin", "win32"):
-    from DisplayCAL.defaultpaths import XDG_CONFIG_DIRS, XDG_CONFIG_HOME
-    from DisplayCAL.edid import get_edid
-    from DisplayCAL.util_x import get_display
-
-    try:
-        from DisplayCAL import xrandr
-    except ImportError:
-        xrandr = None
-    from DisplayCAL.util_os import dlopen, which
-elif sys.platform == "win32":
-    from DisplayCAL import util_win
-    from DisplayCAL.mscms import WCSManagerProxy
-
-    # WCS only available under Vista and later
-    mscms = None if sys.getwindowsversion() < (6,) else WCSManagerProxy()
-
 
 if TYPE_CHECKING:
     import multiprocessing
@@ -130,14 +63,11 @@ from DisplayCAL.icc_profile.codecs import (
 from DisplayCAL.icc_profile.constants import (
     CIIS,
     CMMS,
-    COLOR_PROFILE_SUBTYPE,
-    COLOR_PROFILE_TYPE,
     DEBUG,
     ENC,
     ENCODINGS,
     ERROR_PROFILE_NOT_ASSOCIATED_WITH_DEVICE,
     ERROR_SUCCESS,
-    FS_ENC,
     GAMUT_VOLUME_ADOBERGB,
     GAMUT_VOLUME_SMPTE431_P3,
     GAMUT_VOLUME_SRGB,
@@ -146,7 +76,15 @@ from DisplayCAL.icc_profile.constants import (
     PROFILE_CLASS,
     TAGS,
     TECH,
-    WCS_PROFILE_MANAGEMENT_SCOPE,
+)
+from DisplayCAL.icc_profile.display_profile import (
+    _winreg_get_display_profiles,
+    get_display_profile,
+    get_display_profile_linux,
+    get_display_profile_macos,
+    get_display_profile_windows,
+    set_display_profile,
+    unset_display_profile,
 )
 from DisplayCAL.icc_profile.structures import (
     ADict,
@@ -1880,788 +1818,6 @@ def create_synthetic_hlg_clut_profile(
         logfile,
         cat,
     )
-
-
-def _colord_get_display_profile(
-    display_no: int = 0, path_only: bool = False, use_cache: bool = True
-) -> None | str | ICCProfile:
-    """Use a brute force way of getting display profile.
-
-    Args:
-        display_no (int): The display number to query.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-        use_cache (bool, optional): If True, use cached profile if available.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    edid_ = get_edid(display_no)
-    device_ids = []
-    if edid_:
-        # Try a range of possible device IDs
-        dife = colord.device_id_from_edid
-        device_ids = [
-            dife(edid_, quirk=False, query=True),
-            dife(edid_, quirk=True, truncate_edid_strings=True),
-            dife(edid_, quirk=True, use_serial_32=False),
-            dife(edid_, quirk=True, use_serial_32=False, truncate_edid_strings=True),
-            dife(edid_, quirk=True),
-            dife(edid_, quirk=False, truncate_edid_strings=True),
-            dife(edid_, quirk=False, use_serial_32=False),
-            dife(edid_, quirk=False, use_serial_32=False, truncate_edid_strings=True),
-            # Try with manufacturer omitted
-            dife(edid_, omit_manufacturer=True),
-            dife(edid_, truncate_edid_strings=True, omit_manufacturer=True),
-            dife(edid_, use_serial_32=False, omit_manufacturer=True),
-            dife(
-                edid_,
-                use_serial_32=False,
-                truncate_edid_strings=True,
-                omit_manufacturer=True,
-            ),
-        ]
-    else:
-        # Fall back to XrandR name
-        try:
-            from DisplayCAL import real_display_size_mm
-        except ImportError as exception:
-            warnings.warn(str(exception), Warning, stacklevel=2)
-            return None
-        display = real_display_size_mm.get_display(display_no)
-        if display:
-            xrandr_name = display.get("xrandr_name")
-            if xrandr_name:
-                edid_ = {"monitor_name": xrandr_name}
-                device_ids = [f"xrandr-{xrandr_name.decode()}"]
-            elif os.getenv("XDG_SESSION_TYPE") == "wayland":
-                # Preliminary Wayland support under non-GNOME desktops.
-                # This still needs a lot of work.
-                device_ids = colord.get_display_device_ids()
-                if device_ids and display_no < len(device_ids):
-                    edid_ = {
-                        "monitor_name": device_ids[display_no].split("xrandr-", 1).pop()
-                    }
-                    device_ids = [device_ids[display_no]]
-    if not edid_:
-        return None
-    for device_id in dict.fromkeys(device_ids):
-        if not device_id:
-            continue
-        try:
-            profile = colord.get_default_profile(device_id)
-            profile_path = profile.properties.get("Filename")
-        except colord.CDObjectQueryError:
-            # Device ID was not found, try next one
-            continue
-        except colord.CDError as exception:
-            warnings.warn(str(exception), Warning, stacklevel=2)
-        except colord.DBusException as exception:
-            warnings.warn(str(exception), Warning, stacklevel=2)
-        else:
-            if profile_path:
-                if "hash" in edid_:
-                    colord.device_ids[edid_["hash"]] = device_id
-                if path_only:
-                    print(
-                        "Got profile from colord for display "
-                        f"{int(display_no):d} ({device_id}):",
-                        profile_path,
-                    )
-                    return profile_path
-                return ICCProfile(profile_path, use_cache=use_cache)
-        break
-    return None
-
-
-def _ucmm_get_display_profile(
-    display_no: int, name: str | bytes, path_only: bool = False, use_cache: bool = True
-) -> None | str | ICCProfile:
-    """Argyll UCMM.
-
-    Args:
-        display_no (int): The display number to query.
-        name (str | bytes): The display name to search for.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-        use_cache (bool, optional): If True, use cached profile if available.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    search = []
-    edid = get_edid(display_no)
-    if edid:
-        # Look for matching EDID entry first
-        search.append((b"EDID", b"0x" + binascii.hexlify(edid["edid"]).upper()))
-    # Fallback to X11 name
-    search.append((b"NAME", name))
-    for path in [XDG_CONFIG_HOME, *XDG_CONFIG_DIRS]:
-        color_jcnf = os.path.join(path, "color.jcnf")
-        if not os.path.isfile(color_jcnf):
-            continue
-
-        with open(color_jcnf) as f:
-            data = json.load(f)
-        displays = data.get("devices", {}).get("display")
-        if not isinstance(displays, dict):
-            continue
-
-        # Look for matching entry
-        for key, value in search:
-            for item in displays.values():
-                if not isinstance(item, dict):
-                    continue
-                if item.get(key) != value:
-                    continue
-                profile_path = item.get("ICC_PROFILE")
-                if path_only:
-                    print(
-                        "Got profile from Argyll UCMM for display "
-                        f"{int(display_no):d} ({key} {value}):",
-                        profile_path,
-                    )
-                    return profile_path
-                return ICCProfile(profile_path, use_cache=use_cache)
-    return None
-
-
-def _wcs_get_display_profile(
-    devicekey: str,
-    scope: int = WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"],
-    profile_type: int = COLOR_PROFILE_TYPE["ICC"],
-    profile_subtype: int = COLOR_PROFILE_SUBTYPE["NONE"],
-    profile_id: int = 0,
-    path_only: bool = False,
-    use_cache: bool = True,
-) -> None | str | ICCProfile:
-    """Get display profile using WCS API.
-
-    Args:
-        devicekey (str): The device key to query.
-        scope (int, optional): The scope of the profile management.
-        profile_type (int, optional): The type of the color profile.
-        profile_subtype (int, optional): The subtype of the color profile.
-        profile_id (int, optional): The ID of the color profile.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-        use_cache (bool, optional): If True, use cached profile if available.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    prof = mscms.get_default_color_profile(
-        scope, devicekey, profile_type, profile_subtype, profile_id
-    )
-    if prof:
-        if path_only:
-            return os.path.join(ICCPROFILES[0], prof)
-        return ICCProfile(prof, use_cache=use_cache)
-    return None
-
-
-def _winreg_get_display_profile(
-    monkey: list,
-    current_user: bool = False,
-    path_only: bool = False,
-    use_cache: bool = True,
-    advanced_color_active: bool | None = None,
-) -> None | str | ICCProfile:
-    """Get display profile from Windows registry.
-
-    Args:
-        monkey (list): Registry key path components for the display.
-        current_user (bool): If True, use HKEY_CURRENT_USER, otherwise
-            HKEY_LOCAL_MACHINE.
-        path_only (bool): If True, return the profile path as a string,
-            otherwise return an ICCProfile object.
-        use_cache (bool): If True, use cached profile if available.
-        advanced_color_active (bool | None): Whether the display is currently
-            in Advanced Color (HDR) mode.  If explicitly False, the
-            ICMProfileAC (Windows HDR Calibration) profile is excluded so the
-            SDR profile is returned instead.  None means unknown (include all).
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    filename = None
-    filenames = _winreg_get_display_profiles(
-        monkey,
-        current_user,
-        exclude_advanced_color=advanced_color_active is False,
-    )
-    if filenames:
-        # last existing file in the list is active
-        filename = filenames.pop()
-    if not filename and not current_user:
-        # fall back to sRGB
-        filename = os.path.join(ICCPROFILES[0], "sRGB Color Space Profile.icm")
-    if filename:
-        if path_only:
-            return os.path.join(ICCPROFILES[0], filename)
-        return ICCProfile(filename, use_cache=use_cache)
-    return None
-
-
-def _winreg_get_display_profiles(
-    monkey: list,
-    current_user: bool = False,
-    exclude_advanced_color: bool = False,
-) -> list:
-    """Get display profile filenames from Windows registry.
-
-    Args:
-        monkey (list): Registry key path components for the display.
-        current_user (bool): If True, use HKEY_CURRENT_USER, otherwise
-            HKEY_LOCAL_MACHINE.
-        exclude_advanced_color (bool): If True, skip profiles stored under
-            ICMProfileAC (the Windows HDR Calibration / Advanced Color profile
-            slot).  Pass True when the display is not currently in HDR mode so
-            the SDR profile is used instead.
-
-    Returns:
-        list: List of profile filenames.
-    """
-    filenames = []
-    try:
-        if current_user and sys.getwindowsversion() >= (6,):
-            # Vista / Windows 7 ONLY
-            # User has to place a check in 'use my settings for this device'
-            # in the color management control panel at least once to cause
-            # this key to be created, otherwise it won't exist
-            subkey = "\\".join(
-                [
-                    "Software",
-                    "Microsoft",
-                    "Windows NT",
-                    "CurrentVersion",
-                    "ICM",
-                    "ProfileAssociations",
-                    "Display",
-                    *monkey,
-                ]
-            )
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, subkey)
-        else:
-            subkey = "\\".join(
-                ["SYSTEM", "CurrentControlSet", "Control", "Class", *monkey]
-            )
-            key = winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE, subkey)
-        numsubkeys, numvalues, mtime = winreg.QueryInfoKey(key)
-        for i in range(numvalues):
-            name, value, type_ = winreg.EnumValue(key, i)
-            if name not in ["ICMProfile", "ICMProfileAC"] or not value:
-                continue
-            if name == "ICMProfileAC" and exclude_advanced_color:
-                continue
-
-            if type_ == winreg.REG_BINARY:
-                # Win2k/XP
-                # convert to list of strings
-                value = value.decode("utf-16").split("\0")
-            elif type_ == winreg.REG_MULTI_SZ:
-                # Vista / Windows 7
-                # nothing to be done, _winreg returns a list of strings
-                pass
-            if not isinstance(value, list):
-                value = [value]
-            while "" in value:
-                value.remove("")
-            filenames.extend(value)
-        winreg.CloseKey(key)
-    except OSError as exception:
-        if exception.args[0] == 2:
-            # Key does not exist
-            pass
-        else:
-            raise
-    return [
-        filename
-        for filename in filenames
-        if os.path.isfile(os.path.join(ICCPROFILES[0], filename))
-    ]
-
-
-def get_display_profile(
-    display_no: int = 0,
-    x_hostname: None | str = None,
-    x_display: None | str = None,
-    x_screen: None | int = None,
-    path_only: bool = False,
-    devicekey: None | str = None,
-    use_active_display_device: bool = True,
-    use_registry: bool = True,
-) -> None | str | ICCProfile:
-    """Return ICC Profile for display n or None.
-
-    Args:
-        display_no (int, optional): The display number to query. Defaults to 0.
-        x_hostname (str, optional): The X server hostname.
-        x_display (str, optional): The X display name.
-        x_screen (int, optional): The X screen number.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-        devicekey (None | str, optional): The device key to query. If None,
-            the active display device will be used.
-        use_active_display_device (bool, optional): If True, use the active
-            display device, otherwise use the first display device.
-        use_registry (bool, optional): If True, use the Windows registry to
-            get the display profile.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    if sys.platform == "win32":
-        return get_display_profile_windows(
-            display_no, path_only, devicekey, use_active_display_device, use_registry
-        )
-    if sys.platform == "darwin":
-        return get_display_profile_macos(display_no, path_only)
-    return get_display_profile_linux(
-        display_no, x_hostname, x_display, x_screen, path_only
-    )
-
-
-def get_display_profile_windows(
-    display_no: int = 0,
-    path_only: bool = False,
-    devicekey: None | str = None,
-    use_active_display_device: bool = True,
-    use_registry: bool = True,
-) -> None | str | ICCProfile:
-    """Return ICC Profile for the given display under Windows.
-
-    Args:
-        display_no (int): The display number to query.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-        devicekey (None | str, optional): The device key to query. If None, the
-            active display device will be used.
-        use_active_display_device (bool, optional): If True, use the active
-            display device, otherwise use the first display device.
-        use_registry (bool, optional): If True, use the Windows registry to
-            get the display profile.
-
-    Raises:
-        ImportError: If pywin32 is not available.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    profile = None
-    if "win32api" not in sys.modules:
-        raise ImportError("pywin32 not available")
-    gdi_device_name = None
-    if not devicekey:
-        # The ordering will work as long as Argyll continues using
-        # EnumDisplayMonitors
-        monitors = util_win.get_real_display_devices_info()
-        moninfo = monitors[display_no]
-        gdi_device_name = moninfo["Device"]
-    if not mscms and not devicekey:
-        # Via GetICMProfile. Sucks royally in a multi-monitor setup
-        # where one monitor is disabled, because it'll always get
-        # the profile of the first monitor regardless if that is the active
-        # one or not. Yuck. Also, in this case it does not reflect runtime
-        # changes to profile assignments. Double yuck.
-        buflen = ctypes.c_ulong(260)
-        dc = win32gui.CreateDC(moninfo["Device"], None, None)
-        try:
-            buf = ctypes.create_unicode_buffer(buflen.value)
-            if ctypes.windll.gdi32.GetICMProfileW(
-                dc,
-                ctypes.byref(buflen),
-                ctypes.byref(buf),  # WCHARs
-            ):
-                if path_only:
-                    profile = buf.value
-                else:
-                    profile = ICCProfile(buf.value, use_cache=True)
-        finally:
-            win32gui.DeleteDC(dc)
-    else:
-        if devicekey:
-            device = None
-        elif use_active_display_device:
-            # This would be the correct way. Unfortunately that is not
-            # what other apps (or Windows itself) do.
-            device = util_win.get_active_display_device(moninfo["Device"])
-        else:
-            # This is wrong, but it's what other apps use. Matches
-            # GetICMProfile sucky behavior i.e. should return the same
-            # profile, but atleast reflects runtime changes to profile
-            # assignments.
-            device = util_win.get_first_display_device(moninfo["Device"])
-        if device:
-            devicekey = device.DeviceKey
-    if devicekey:
-        if mscms:
-            # Via WCS
-            if util_win.per_user_profiles_isenabled(devicekey=devicekey):
-                scope = WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"]
-            else:
-                scope = WCS_PROFILE_MANAGEMENT_SCOPE["SYSTEM_WIDE"]
-            if not use_registry:
-                # NOTE: WcsGetDefaultColorProfile causes the whole system
-                # to hitch if the profile of the active display device is
-                # queried. Windows bug?
-                return _wcs_get_display_profile(
-                    str(devicekey), scope, path_only=path_only
-                )
-        else:
-            scope = None
-            # Via registry
-        monkey = devicekey.split("\\")[-2:]  # pun totally intended
-        # Current user scope
-        current_user = scope == WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"]
-        # Detect Advanced Color (HDR) state so we can skip the ICMProfileAC
-        # entry when the display is in SDR mode (issue #627).
-        if gdi_device_name is None:
-            gdi_device_name = util_win.get_gdi_device_name_for_devicekey(devicekey)
-        advanced_color_active = (
-            util_win.is_advanced_color_enabled(gdi_device_name)
-            if gdi_device_name
-            else None
-        )
-        if current_user:
-            profile = _winreg_get_display_profile(
-                monkey,
-                True,
-                path_only=path_only,
-                advanced_color_active=advanced_color_active,
-            )
-        else:
-            # System scope
-            profile = _winreg_get_display_profile(
-                monkey,
-                path_only=path_only,
-                advanced_color_active=advanced_color_active,
-            )
-
-    return profile
-
-
-def get_display_profile_macos(
-    display_no: int = 0,
-    path_only: bool = False,
-) -> None | str | ICCProfile:
-    """Return ICC Profile for the given display under macOS.
-
-    Args:
-        display_no (int, optional): The display number to query. Defaults to 0.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-
-    Raises:
-        OSError: If there is an error executing the AppleScript command.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    from DisplayCAL.util_mac import osascript
-
-    if intlist(platform.mac_ver()[0].split(".")) >= [10, 6]:
-        options = ["Image Events"]
-    else:
-        options = ["ColorSyncScripting"]
-
-    for option in options:
-        # applescript: one-based index
-        applescript = [
-            f'tell app "{option}"',
-            "set displayProfile to location of display profile of "
-            f"display {int(display_no + 1):d}",
-            "return POSIX path of displayProfile",
-            "end tell",
-        ]
-        retcode, output, errors = osascript(applescript)
-        if retcode == 0 and output.strip():
-            filename = output.strip("\n").decode(FS_ENC)
-            profile = filename if path_only else ICCProfile(filename, use_cache=True)
-        elif errors.strip():
-            raise OSError(errors.strip())
-
-    return profile
-
-
-def get_display_profile_linux(
-    display_no: int = 0,
-    x_hostname: None | str = None,
-    x_display: None | int = None,
-    x_screen: None | int = None,
-    path_only: bool = False,
-) -> None | str | ICCProfile:
-    """Return ICC Profile for the given display under Linux.
-
-    Args:
-        display_no (int): The display number to query.
-        x_hostname (str, optional): The X server hostname.
-        x_display (int, optional): The X display number.
-        x_screen (int, optional): The X screen number.
-        path_only (bool, optional): If True, return the profile path as a
-            string, otherwise return an ICCProfile object.
-
-    Returns:
-        None | str | ICCProfile: The display profile path as a string or
-            an ICCProfile object, or None if no profile is found.
-    """
-    options = ["_ICC_PROFILE"]
-    try:
-        from DisplayCAL import real_display_size_mm
-    except ImportError as exception:
-        warnings.warn(str(exception), Warning, stacklevel=2)
-        display = get_display()
-    else:
-        display = real_display_size_mm.get_x_display(display_no)
-    if display:
-        if x_hostname is None:
-            x_hostname = display[0]
-        if x_display is None:
-            x_display = display[1]
-        if x_screen is None:
-            x_screen = display[2]
-        x_display_name = f"{x_hostname}:{x_display}.{x_screen}"
-    for option in options:
-        # Linux
-        # Try colord
-        if colord.which("colormgr") and (
-            profile := (_colord_get_display_profile(display_no, path_only=path_only))
-        ):
-            return profile
-        if path_only:
-            # No way to figure out the profile path from X atom, so use
-            # Argyll's UCMM if libcolordcompat.so is not present
-            if dlopen("libcolordcompat.so"):
-                # UCMM configuration might be stale, ignore
-                return None
-            return _ucmm_get_display_profile(display_no, x_display_name, path_only)
-        # Try XrandR
-        if (
-            xrandr
-            and real_display_size_mm
-            and option == "_ICC_PROFILE"
-            and None not in (x_hostname, x_display, x_screen)
-        ):
-            with xrandr.XDisplay(x_display_name) as display:
-                if DEBUG:
-                    print("Using XrandR")
-                for i, atom_id in enumerate(
-                    [
-                        real_display_size_mm.get_x_icc_profile_output_atom_id(
-                            display_no
-                        ),
-                        real_display_size_mm.get_x_icc_profile_atom_id(display_no),
-                    ]
-                ):
-                    if not atom_id:
-                        continue
-                    if i == 0:
-                        meth = display.get_output_property
-                        what = real_display_size_mm.get_xrandr_output_xid(display_no)
-                    else:
-                        meth = display.get_window_property
-                        what = display.root_window(0)
-                    try:
-                        window_property = meth(what, atom_id)
-                    except ValueError as exception:
-                        warnings.warn(str(exception), Warning, stacklevel=2)
-                    else:
-                        if window_property and (
-                            profile := ICCProfile(
-                                b"".join(
-                                    bytes(chr(n), "utf-8") for n in window_property
-                                ),
-                                use_cache=True,
-                            )
-                        ):
-                            return profile
-                    if DEBUG:
-                        if i == 0:
-                            print("Couldn't get _ICC_PROFILE XrandR output property")
-                            print("Using X11")
-                        else:
-                            print("Couldn't get _ICC_PROFILE X atom")
-            return None
-
-        # Read up to 8 MB of any X properties
-        if DEBUG:
-            print("Using xprop")
-        xprop = which("xprop")
-        if not xprop:
-            return None
-        atom = "{}{}".format(option, "" if display_no == 0 else f"_{display_no}")
-        tgt_proc = sp.Popen(
-            [
-                xprop,
-                "-display",
-                f"{x_hostname}:{x_display}.{x_screen}",
-                "-len",
-                "8388608",
-                "-root",
-                "-notype",
-                atom,
-            ],
-            stdin=sp.PIPE,
-            stdout=sp.PIPE,
-            stderr=sp.PIPE,
-        )
-        stdout, stderr = [data.strip(b"\n") for data in tgt_proc.communicate()]
-        if stdout:
-            raw = [item.strip() for item in stdout.split("=")]
-            if raw[0] == atom and len(raw) == 2:
-                binary_data = "".join([chr(int(part)) for part in raw[1].split(", ")])
-                profile = ICCProfile(binary_data, use_cache=True)
-        elif stderr and tgt_proc.wait() != 0:
-            raise OSError(stderr)
-        if profile:
-            break
-    return profile
-
-
-def _wcs_set_display_profile(
-    devicekey: str,
-    profile_name: str,
-    scope: int = WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"],
-) -> bool:
-    """Set the current default WCS color profile for the given device.
-
-    If the device is a display, this will also set its video card gamma ramps
-    to linear* if the given profile is the display's current default profile
-    and Windows calibration management isn't enabled.
-
-    Note that the profile needs to have been already installed.
-
-    * 0..65535 will get mapped to 0..65280, which is a Windows bug.
-
-    Args:
-        devicekey (str): The device key of the display.
-        profile_name (str): The name of the profile to be set.
-        scope (int): The scope of the profile management, either
-            WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"] or
-            WCS_PROFILE_MANAGEMENT_SCOPE["SYSTEM_WIDE"].
-
-    Returns:
-        bool: True if the profile was set successfully, False otherwise.
-    """
-    mscms.associate_color_profile_with_device(scope, profile_name, str(devicekey))
-    profiles = mscms.get_device_color_profile_list(scope, str(devicekey))
-    return profile_name in profiles
-
-
-def _wcs_unset_display_profile(
-    devicekey: str,
-    profile_name: str,
-    scope: int = WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"],
-) -> bool:
-    """Unset the current default WCS color profile for the given device.
-
-    If the device is a display, this will also set its video card gamma ramps
-    to linear* if the given profile is the display's current default profile
-    and Windows calibration management isn't enabled.
-
-    Note that the profile needs to have been already installed.
-
-    * 0..65535 will get mapped to 0..65280, which is a Windows bug.
-
-    Args:
-        devicekey (str): The device key of the display.
-        profile_name (str): The name of the profile to be unset.
-        scope (int): The scope of the profile management, either
-            WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"] or
-            WCS_PROFILE_MANAGEMENT_SCOPE["SYSTEM_WIDE"].
-
-    Returns:
-        bool: True if the profile was unset successfully, False otherwise.
-    """
-    mscms.disassociate_color_profile_from_device(scope, profile_name, str(devicekey))
-    profiles = mscms.get_device_color_profile_list(scope, str(devicekey))
-    return profile_name not in profiles
-
-
-def set_display_profile(
-    profile_name: str,
-    display_no: int = 0,
-    devicekey: None | str = None,
-    use_active_display_device: bool = True,
-) -> bool:
-    """Set the current default WCS color profile for the given device.
-
-    Args:
-        profile_name (str): The name of the profile to be set.
-        display_no (int): The display number to set the profile for.
-        devicekey (str): The device key of the display.
-        use_active_display_device (bool): Whether to use the active display
-            device.
-
-    Returns:
-        bool: True if the profile was set successfully, False otherwise.
-    """
-    # Currently only implemented for Windows.
-    # The profile to be assigned has to be already installed!
-    if not devicekey:
-        device = util_win.get_display_device(display_no, use_active_display_device)
-        if not device:
-            return False
-        devicekey = device.DeviceKey
-    if mscms:
-        if util_win.per_user_profiles_isenabled(devicekey=devicekey):
-            scope = WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"]
-        else:
-            scope = WCS_PROFILE_MANAGEMENT_SCOPE["SYSTEM_WIDE"]
-        return _wcs_set_display_profile(str(devicekey), profile_name, scope)
-    # TODO: Implement for XP
-    return False
-
-
-def unset_display_profile(
-    profile_name: str,
-    display_no: int = 0,
-    devicekey: None | str = None,
-    use_active_display_device: bool = True,
-) -> bool:
-    """Unset the current default WCS color profile for the given device.
-
-    If the device is a display, this will also set its video card gamma ramps
-    to linear* if the given profile is the display's current default profile
-    and Windows calibration management isn't enabled.
-
-    Note that the profile needs to have been already installed.
-    * 0..65535 will get mapped to 0..65280, which is a Windows bug.
-
-    Args:
-        profile_name (str): The name of the profile to be unset.
-        display_no (int, optional): The display number to unset the profile
-            for. Defaults to 0.
-        devicekey (None | str): The device key of the display. Defatults to
-            None, which means the active display device will be used.
-        use_active_display_device (bool, optional): Whether to use the active
-            display device. Defaults to True.
-
-    Returns:
-        bool: True if the profile was unset successfully, False otherwise.
-    """
-    # Currently only implemented for Windows.
-    # The profile to be unassigned has to be already installed!
-    if not devicekey:
-        device = util_win.get_display_device(display_no, use_active_display_device)
-        if not device:
-            return False
-        devicekey = device.DeviceKey
-    if mscms:
-        if util_win.per_user_profiles_isenabled(devicekey=devicekey):
-            scope = WCS_PROFILE_MANAGEMENT_SCOPE["CURRENT_USER"]
-        else:
-            scope = WCS_PROFILE_MANAGEMENT_SCOPE["SYSTEM_WIDE"]
-        return _wcs_unset_display_profile(str(devicekey), profile_name, scope)
-    # TODO: Implement for XP
-    return False
 
 
 def _blend_blackpoint(
